0506 sync-results PG 타입추론 에러 긴급 수정 — sent_count = (::int + ::int) 명시 캐스팅 (operator is not unique unknown+unknown 해결)

+++ packages/backend/src/utils/campaign-lifecycle.ts
@@ -247,7 +247,7 @@ export async function syncCampaignResults(companyId: string): Promise<SyncResult
         if (runInfo.rows.length > 0) {
           await query(
             `UPDATE campaigns SET
-              sent_count = $1 + $2,
+              sent_count = ($1::int + $2::int),
               success_count = $1,
               fail_count = $2,
               status = $3::text,
@@ -355,7 +355,7 @@ export async function syncCampaignResults(companyId: string): Promise<SyncResult
 
         await query(
           `UPDATE campaigns SET
-            sent_count = $1 + $2,
+            sent_count = ($1::int + $2::int),
             success_count = $1,
             fail_count = $2,
             status = $3::text,
